+++ gpjax/variational.py
@@ -67,53 +67,3 @@ class VariationalGaussian(VariationalFamily):
         }
         return hyperparams
 
-
-@dataclass
-class VariationalPosterior:
-    posterior: Posterior
-    variational_family: VariationalFamily
-    jitter: Optional[float] = DEFAULT_JITTER
-
-    def __post_init__(self):
-        self.prior = self.posterior.prior
-        self.likelihood = self.posterior.likelihood
-
-    @property
-    def params(self) -> Dict:
-        hyperparams = concat_dictionaries(
-            self.posterior.params,
-            {"variational_family": self.variational_family.params},
-        )
-        return hyperparams
-
-    @abc.abstractmethod
-    def mean(self, train_data: Dataset, params: dict) -> Callable[[Dataset], Array]:
-        raise NotImplementedError
-
-    @abc.abstractmethod
-    def variance(self, train_data: Dataset, params: dict) -> Callable[[Dataset], Array]:
-        raise NotImplementedError
-
-    @abc.abstractmethod
-    def elbo(
-        self, train_data: Dataset, transformations: Dict
-    ) -> Callable[[Array], Array]:
-        raise NotImplementedError
-
-
-def ELBO(
-    model: VariationalPosterior, train_data: Dataset, transformations: Dict
-) -> Callable[[Array], Array]:
-    def elbo(*args, **kwargs):
-        return model.elbo(train_data, transformations)(*args, **kwargs)
-
-    return elbo
-
-
-def VFE(
-    model: VariationalPosterior, train_data: Dataset, transformations: Dict
-) -> Callable[[Array], Array]:
-    def vfe(*args, **kwargs):
-        return -model.elbo(train_data, transformations)(*args, **kwargs)
-
-    return vfe
